favorite auth inprocess

+++ controllers/contactsControllers.js
@@ -11,7 +11,7 @@ import mongoose from "mongoose";
 // GET /api/contacts
 export const getAllContacts = async (req, res, next) => {
   try {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({ owner: req.user.id }); // передамо в параметри фільтр - об'єкт з полем id користувача
     return res.status(200).json(contacts);
   } catch (error) {
     console.error(error);
@@ -32,6 +32,11 @@ export const getOneContact = async (req, res, next) => {
     if (!contact) {
       throw HttpError(404);
     }
+    // перевірка належності контакта користувачу
+    if (contact.owner.toString() !== req.user.id) {
+      throw HttpError(403, "Contact not found");
+    }
+
     res.status(200).json(contact);
   } catch (error) {
     const status = error.status || 500; // обробка error як що статус відсутний
@@ -47,6 +52,16 @@ export const deleteContact = async (req, res, next) => {
     if (!mongoose.Types.ObjectId.isValid(id)) {
       throw HttpError(400, "Invalid ObjectId format");
     }
+    // пошук контакту за його ідентифікатором
+    const contact = await Contact.findById(id);
+    if (!contact) {
+      throw HttpError(404);
+    }
+    // перевірка належності контакта користувачу
+    if (contact.owner.toString() !== req.user.id) {
+      throw HttpError(403, "Contact not found");
+    }
+    // пошук та видалення котакту
     const removedContact = await Contact.findByIdAndDelete(id);
     if (!removedContact) {
       throw HttpError(404);
@@ -71,7 +86,7 @@ export const createContact = async (req, res, next) => {
       name,
       email,
       phone,
-      ownerId: req.user.id, //додамо id користувача
+      owner: req.user.id, // add owner з моделі Contact та присвоєння їй id користувача
     });
 
     const savedContact = await newContact.save();
@@ -93,6 +108,15 @@ export const updateContact = async (req, res, next) => {
     const { error } = updateContactSchema.validate(req.body);
     if (error) {
       throw HttpError(400, "Body must have at least one field");
+    }
+    // пошук контакту за його ідентифікатором
+    const contact = await Contact.findById(id);
+    if (!contact) {
+      throw HttpError(404);
+    }
+    // перевірка належності контакта користувачу
+    if (contact.owner.toString() !== req.user.id) {
+      throw HttpError(403, "Contact not found");
     }
     const updatedContact = await Contact.findByIdAndUpdate(
       id,
@@ -147,6 +171,19 @@ export const updateContactFavoriteStatus = async (req, res, next) => {
     // details[0].message - перша помилка при перевірці тіла запиту
     return res.status(400).json({ message: error.details[0].message });
   }
+
+  // пошук контакту за його ідентифікатором
+  const contact = await Contact.findById(contactId);
+  console.log(contact._id);
+  if (!contact) {
+    throw HttpError(404);
+    return res.status(404).json({ message: error.message });
+  }
+  // перевірка належності контакта користувачу
+  if (contact.owner.toString() !== req.user.id) {
+    throw HttpError(403, "Contact not found");
+  }
+
   // Якщо з body все добре, виклик ф-ції updateStatusContact (contactId, body)
   try {
     //перевірка contactId на належність до ObjectId
@@ -161,4 +198,13 @@ export const updateContactFavoriteStatus = async (req, res, next) => {
   } catch (error) {
     next(error);
   }
+};
+export default {
+  getAllContacts,
+  getOneContact,
+  deleteContact,
+  createContact,
+  updateContact,
+  updateStatusContact,
+  updateContactFavoriteStatus,
 };
